[REFACTOR/#41] pages 레이어 각 폴더 최상위  public api 제거
