+++ src/app/(auth)/login.tsx
@@ -1,4 +1,4 @@
-import { LoginPage } from "@/pages/auth";
+import { LoginPage } from "@/pages/auth/login";
 
 export default function LoginScreen() {
 	return <LoginPage />;
