+++ src/app/(auth)/register.tsx
@@ -1,4 +1,4 @@
-import { RegisterPage } from "@/pages/auth";
+import { RegisterPage } from "@/pages/auth/register";
 
 export default function RegisterScreen() {
 	return <RegisterPage />;
