+++ src/app/(protected)/(admin)/chat.tsx
@@ -1,4 +1,4 @@
-import { AdminChatPage } from "@/pages/admin";
+import { AdminChatPage } from "@/pages/admin/chat";
 
 export default function AdminChatScreen() {
 	return <AdminChatPage />;
