+++ src/app/(protected)/(admin)/dashboard.tsx
@@ -1,4 +1,4 @@
-import { AdminDashboardPage } from "@/pages/admin";
+import { AdminDashboardPage } from "@/pages/admin/dashboard";
 
 export default function AdminDashboardScreen() {
 	return <AdminDashboardPage />;
