+++ src/app/(protected)/(admin)/home.tsx
@@ -1,4 +1,4 @@
-import { AdminHomePage } from "@/pages/admin";
+import { AdminHomePage } from "@/pages/admin/home";
 
 export default function AdminHomeScreen() {
 	return <AdminHomePage />;
