+++ src/app/(protected)/(admin)/profile.tsx
@@ -1,4 +1,4 @@
-import { AdminProfilePage } from "@/pages/admin";
+import { AdminProfilePage } from "@/pages/admin/profile";
 
 export default function AdminProfileScreen() {
 	return <AdminProfilePage />;
