+++ src/app/(protected)/(partner)/dashboard.tsx
@@ -1,4 +1,4 @@
-import { PartnerDashboardPage } from "@/pages/partner";
+import { PartnerDashboardPage } from "@/pages/partner/dashboard";
 
 export default function PartnerDashboardScreen() {
 	return <PartnerDashboardPage />;
