+++ src/app/(protected)/(partner)/map.tsx
@@ -1,4 +1,4 @@
-import { PartnerMapPage } from "@/pages/partner";
+import { PartnerMapPage } from "@/pages/partner/map";
 
 export default function PartnerMapScreen() {
 	return <PartnerMapPage />;
