+++ src/app/(protected)/(partner)/profile.tsx
@@ -1,4 +1,4 @@
-import { PartnerProfilePage } from "@/pages/partner";
+import { PartnerProfilePage } from "@/pages/partner/profile";
 
 export default function PartnerProfileScreen() {
 	return <PartnerProfilePage />;
