+++ src/app/(protected)/(student)/home.tsx
@@ -1,4 +1,4 @@
-import { StudentHomePage } from "@/pages/student";
+import { StudentHomePage } from "@/pages/student/home";
 
 export default function StudentHomeScreen() {
 	return <StudentHomePage />;
