+++ src/app/(protected)/(student)/profile.tsx
@@ -1,4 +1,4 @@
-import { StudentProfilePage } from "@/pages/student";
+import { StudentProfilePage } from "@/pages/student/profile";
 
 export default function StudentProfileScreen() {
 	return <StudentProfilePage />;
