+++ src/app/(protected)/(student)/suggestion.tsx
@@ -1,4 +1,4 @@
-import { StudentSuggestionPage } from "@/pages/student";
+import { StudentSuggestionPage } from "@/pages/student/suggestion";
 
 export default function StudentSuggestionScreen() {
 	return <StudentSuggestionPage />;
