--- src/pages/admin/index.ts
@@ -1,5 +1,0 @@
-export { AdminHomePage } from "./home";
-export { AdminMapPage } from "./map";
-export { AdminDashboardPage } from "./dashboard";
-export { AdminChatPage } from "./chat";
-export { AdminProfilePage } from "./profile";
