--- src/pages/auth/index.ts
@@ -1,2 +1,0 @@
-export { LoginPage } from "./login";
-export { RegisterPage } from "./register";
