+++ src/pages/dev-hub/index.ts
@@ -1,0 +1,1 @@
+export { DevHubPage } from "./ui/DevHubPage";
