--- src/pages/index/index.ts
@@ -1,1 +1,0 @@
-export { IndexPage } from "./ui/IndexPage";
