--- src/pages/partner/index.ts
@@ -1,5 +1,0 @@
-export { PartnerHomePage } from "./home";
-export { PartnerMapPage } from "./map";
-export { PartnerDashboardPage } from "./dashboard";
-export { PartnerChatPage } from "./chat";
-export { PartnerProfilePage } from "./profile";
