--- src/pages/student/index.ts
@@ -1,4 +1,0 @@
-export { StudentHomePage } from "./home";
-export { StudentMapPage } from "./map";
-export { StudentSuggestionPage } from "./suggestion";
-export { StudentProfilePage } from "./profile";
